Reestruturação projeto + Responsividade OK
